refactor(ai): split customer tools by domain
