+++ apps/server/src/domain/ai/tools/rentals/presenter.ts
@@ -1,13 +1,17 @@
+import type { ReturnSlotRow } from "@/domain/rentals";
+
+import { returnSlotExpiresAt } from "@/domain/rentals";
 import { toContractRental } from "@/http/presenters/rentals.presenter";
 
-import type { CreateCustomerToolsArgs } from "./customer-tool-helpers";
+import type { RentalQueryToolsArgs } from "../shared/customer-tool-args";
 
 import {
   formatLocalDateTime,
   formatMinorVnd,
   getRentalStatusLabel,
-  getStationByIdOrNull,
-} from "./customer-tool-helpers";
+  getReturnSlotStatusLabel,
+} from "../shared/customer-tool-formatters";
+import { getStationByIdOrNull } from "../shared/customer-tool-lookups";
 
 type RentalRow = Parameters<typeof toContractRental>[0];
 
@@ -18,7 +22,7 @@ type StationSummary = {
 };
 
 export async function loadStationSummaryMap(
-  args: CreateCustomerToolsArgs,
+  args: Pick<RentalQueryToolsArgs, "stationQueryService">,
   rentals: readonly RentalRow[],
 ) {
   const stationIds = [...new Set(rentals.flatMap(rental => [rental.startStationId, rental.endStationId].filter((stationId): stationId is string => !!stationId)))];
@@ -71,5 +75,27 @@ export function toRentalDetailItem(
     statusLabel: getRentalStatusLabel(rental.status),
     totalPriceDisplay: formatMinorVnd(rental.totalPrice),
     updatedAtDisplay: formatLocalDateTime(rental.updatedAt),
+  };
+}
+
+export function toReturnSlotAiDetail(
+  returnSlot: ReturnSlotRow,
+  station: { id: string; name: string; address: string } | null,
+) {
+  return {
+    id: returnSlot.id,
+    rentalId: returnSlot.rentalId,
+    userId: returnSlot.userId,
+    stationId: returnSlot.stationId,
+    reservedFrom: returnSlot.reservedFrom.toISOString(),
+    expiresAt: returnSlotExpiresAt(returnSlot.reservedFrom).toISOString(),
+    reservedFromDisplay: formatLocalDateTime(returnSlot.reservedFrom),
+    status: returnSlot.status,
+    statusLabel: getReturnSlotStatusLabel(returnSlot.status),
+    createdAt: returnSlot.createdAt.toISOString(),
+    createdAtDisplay: formatLocalDateTime(returnSlot.createdAt),
+    updatedAt: returnSlot.updatedAt.toISOString(),
+    updatedAtDisplay: formatLocalDateTime(returnSlot.updatedAt),
+    station,
   };
 }
