+++ apps/server/src/domain/ai/tools/shared/customer-tool-schema-common.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+
+export const StationSummaryInfoSchema = z.object({
+  id: z.uuidv7(),
+  name: z.string(),
+  address: z.string(),
+}).strict();
+
+export const ActionFailureKindSchema = z.enum([
+  "validation",
+  "business",
+  "temporary",
+]);
